+++ Data/SBV/Core/Symbolic.hs
@@ -393,6 +393,7 @@ data SetOp = SetEqual
            | SetSubset
            | SetDifference
            | SetComplement
+           | SetHasSize
         deriving (Eq, Ord)
 
 -- The show instance for 'SetOp' is merely for debugging, we map them separately so
@@ -407,6 +408,7 @@ instance Show SetOp where
   show SetSubset     = "Set.subset"
   show SetDifference = "Set.difference"
   show SetComplement = "Set.complement"
+  show SetHasSize    = "Set.setHasSize"
 
 -- Show instance for 'Op'. Note that this is largely for debugging purposes, not used
 -- for being read by any tool.