+++ Data/SBV/Set.hs
@@ -36,14 +36,14 @@ module Data.SBV.Set (
         , insert, delete
 
         -- * Query
-        , member, notMember, null, isEmpty, isFull, isUniversal, isSubsetOf, isProperSubsetOf, disjoint
+        , member, notMember, null, isEmpty, isFull, isUniversal, isFinite, isInfinite, hasSize, isSubsetOf, isProperSubsetOf, disjoint
 
         -- * Combinations
         , union, unions, intersection, intersections, difference, (\\)
 
         ) where
 
-import Prelude hiding (null)
+import Prelude hiding (null, isInfinite)
 
 import Data.Proxy (Proxy(Proxy))
 import qualified Data.Set as Set
@@ -292,6 +292,100 @@ isFull = (.== full)
 -- | Synonym for 'Data.SBV.Set.isFull'.
 isUniversal :: HasKind a => SSet a -> SBool
 isUniversal = isFull
+
+-- | Is the given set finite?
+--
+-- >>> isFinite (empty :: SSet Integer)
+-- True
+--
+-- >>> isFinite (full :: SSet Integer)
+-- False
+--
+-- >>> prove $ \a e -> isFinite (a :: SSet Integer) .=> isFinite (e `insert` a)
+-- Q.E.D.
+--
+-- >>> prove $ \a e -> isFinite (a :: SSet Integer) .=> isFinite (e `delete` a)
+-- Q.E.D.
+--
+-- >>> prove $ \a b -> isFinite (a :: SSet Integer) .&& isFinite b .=> isFinite (a `union` b)
+-- Q.E.D.
+--
+-- >>> prove $ \a b -> isFinite (a :: SSet Integer) .&& isFinite b .=> isFinite (a `intersection` b)
+-- Q.E.D.
+--
+-- >>> prove $ \a -> isFinite (a :: SSet Integer) .<=> isInfinite (complement a)
+-- Q.E.D.
+isFinite :: (Ord a, SymVal a) => SSet a -> SBool
+isFinite sa
+  -- Case 1: Constant regular set: always finite
+  | Just (RegularSet _) <- unliteral sa
+  = sTrue
+
+  -- Case 2: Constant complement set: never finite
+  | Just (ComplementSet _) <- unliteral sa
+  = sFalse
+
+  -- Otherwise, go symbolic with the existential variable
+  | True
+  = SBV $ SVal KBool $ Right $ cache r
+  where r st = do sva <- sbvToSV st sa
+                  k   <- internalVariable st KUnbounded
+
+                  -- Assert that has-size with `k` returns true
+                  fin <- newExpr st KBool $ SBVApp (SetOp SetHasSize) [sva, k]
+
+                  internalConstraint st False [] $ SVal KBool $ Right $ cache $ \_ -> return fin
+
+                  -- we're good to go
+                  return fin
+
+-- | Is the given set infinite?
+--
+-- >>> prove $ \a e -> isInfinite (a :: SSet Integer) .=> isInfinite (e `delete` a)
+-- Q.E.D.
+--
+-- >>> prove $ \a e -> isInfinite (a :: SSet Integer) .=> isInfinite (e `insert` a)
+-- Q.E.D.
+isInfinite :: (Ord a, SymVal a) => SSet a -> SBool
+isInfinite = sNot . isFinite
+
+-- | Does the set have the given size? Note that this will
+-- ensure the set under consideration is finite.
+--
+-- >>> prove $ \i -> hasSize (empty :: SSet Integer) i .== (i .== 0)
+-- Q.E.D.
+--
+-- >>> sat $ \i -> hasSize (full :: SSet Integer) i
+-- Unsatisfiable
+--
+-- >>> prove $ \a b i j k -> hasSize (a :: SSet Integer) i .&& hasSize (b :: SSet Integer) j .&& hasSize (a `union` b) k .=> k .>= i `smax` j
+-- Q.E.D.
+--
+-- >>> prove $ \a b i j k -> hasSize (a :: SSet Integer) i .&& hasSize (b :: SSet Integer) j .&& hasSize (a `intersection` b) k .=> k .<= i `smin` j
+-- Q.E.D.
+--
+-- >>> prove $ \a k -> hasSize (a :: SSet Integer) k .=> isFinite a
+-- Q.E.D.
+hasSize :: (Ord a, SymVal a) => SSet a -> SInteger -> SBool
+hasSize sa si
+  -- Case 1: Constant regular set, see if the size matches
+  | Just (RegularSet a) <- unliteral sa
+  = literal (fromIntegral (Set.size a)) .== si
+
+  -- Case 2: Constant complement set, will never have finite size
+  | Just (ComplementSet _) <- unliteral sa
+  = sFalse
+
+  -- Case 3: Integer is constant, and is negative:
+  | Just i <- unliteral si, i < 0
+  = sFalse
+
+  -- Otherwise, go symbolic
+  | True
+  = SBV $ SVal KBool $ Right $ cache r
+  where r st = do sva <- sbvToSV st sa
+                  svi <- sbvToSV st si
+                  newExpr st KBool $ SBVApp (SetOp SetHasSize) [sva, svi]
 
 -- | Subset test.
 --